Add files via upload
Project-Based learning for OpenRocket 
PROJECT GOAL
1. Get familiarizes with the software.
2. Understand how different components and their varying geometry effect the stability of overall rocket.
3. Analysis of mass distribution across the model and its effects.
PROBLEM-STATEMENT 
Design and simulate a subscale model rocket using OpenRocket to achieve a minimum apogee of 1 km, with constraints of ≤15 kg total mass and ≤6 ft structural height. The rocket is modeled with an aluminum body and optimized for aerodynamic efficiency, structural feasibility, and launch performance.